requirements.txt
amqp==1.4.9
anyjson==0.3.3
simplejson==3.17.2
asn1crypto==0.24.0
astroid==2.3.3
autopep8==1.4.4
billiard==3.3.0.23
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
configobj==5.0.6
decorator==4.3.0
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-yasg==1.17.0
idna==2.8
isort==4.3.21
itypes==1.2.0
kombu==3.0.37
mysql-connector-python==8.0.20
numpy==1.18.1
packaging==20.4
pipreqs==0.4.10
pyDes==2.0.1
redis==3.5.3
requests==2.22.0
six==1.12.0
urllib3==1.25.7
yarg==0.1.9
cryptography==2.9.2
dbutils==2.0.1
channels==2.0.1
channels-redis==2.0.2
gunicorn==20.1.0
gevent==1.4.0
pandas==1.1.5
daphne==2.2.5
mysqlclient==1.4.6
interval==1.0.0
